powerup activator
powerups are managed trough the powerup manager. the enum in the top
will contain names for all the available powerups.
the Public void ActivatePowerup(PowerupType type) will do a switch on
the powerup that should be activated. if a powerup is not in there, it
will not work.
powerups will automaticly do a callback to DeactivatePowerup that works
the same but removes the powerup you currently have active.